Components: Move LocaleSuggestions from signup to components folder
