+++ assets/stylesheets/_components.scss
@@ -133,6 +133,7 @@
 @import 'components/input-chrono/style';
 @import 'components/logged-out-form/style';
 @import 'components/language-picker/style';
+@import 'components/locale-suggestions/style';
 @import 'components/main/style';
 @import 'components/marketing-survey/cancel-purchase-form/style';
 @import 'components/mobile-back-to-sidebar/style';
@@ -483,7 +484,6 @@
 @import 'reader/update-notice/style';
 @import 'signup/style';
 @import 'signup/flow-progress-indicator/style';
-@import 'signup/locale-suggestions/style';
 @import 'signup/navigation-link/style';
 @import 'signup/processing-screen/style';
 @import 'signup/step-wrapper/style';
