+++ client/components/locale-suggestions/style.scss
